# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tweet Feeder API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized, Please check your credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign Up",
                "description": "Sign Up by name, email, password, password_confirm, phone, twitter_username",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Sign Up user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirm",
                                    "phone",
                                    "twitter_username"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirm": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "5552123433"
                                    },
                                    "twitter_username": {
                                        "type": "string",
                                        "example": "username123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Input Format response"
                    }
                }
            }
        },
        "/api/validate-user-phone": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validate User Phone",
                "description": "Validate User Phone by Phone Number",
                "operationId": "authValidateUserPhone",
                "requestBody": {
                    "description": "Validate User Phone",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "5552123433"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "A1B2C3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Input Format response"
                    }
                }
            }
        },
        "/api/validate-user-email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validate User Email",
                "description": "Validate User Email",
                "operationId": "authValidateUserEmail",
                "requestBody": {
                    "description": "Validate User Email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "A1B2C3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Input Format response"
                    }
                }
            }
        },
        "/api/tweet-list": {
            "post": {
                "tags": [
                    "tweet-endpoints"
                ],
                "requestBody": {
                    "description": "listing of Tweets",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "twitter_username"
                                ],
                                "properties": {
                                    "twitter_username": {
                                        "type": "string",
                                        "example": "username1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Display a listing of Tweets paginate with size 20."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tweets/{id}": {
            "put": {
                "tags": [
                    "tweet-endpoints"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tweet Update",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "text",
                                        "example": "Tweet content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tweet Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tweets/{id}/publish": {
            "post": {
                "tags": [
                    "tweet-endpoints"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Publish Tweet with remote",
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Tweet is published with remote"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tweets/sync": {
            "get": {
                "tags": [
                    "tweet-endpoints"
                ],
                "responses": {
                    "200": {
                        "description": "Tweets syncronized with remote"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": {}
                    }
                },
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}